re-add hook for inserting externally-defined subclasses of RocketTile
